Removed classes from api/config to redirect the code on the geoview-layers classes
