+++ packages/geoview-core/src/core/utils/config/validation-classes/temporary-entry-config.ts
