+++ packages/geoview-core/src/geo/layer/layer.ts
@@ -22,6 +22,7 @@ import {
   LayerGroupCreatedEvent,
   LayerEntryRegisterInitEvent,
   LayerGVCreatedEvent,
+  LayerConfigCreatedEvent,
 } from '@/geo/layer/geoview-layers/abstract-geoview-layers';
 import {
   MapConfigLayerEntry,
@@ -32,8 +33,8 @@ import {
   layerEntryIsGroupLayer,
   TypeLayerStatus,
   GeoCoreLayerConfig,
-  GroupLayerEntryConfig,
 } from '@/api/config/types/map-schema-types';
+import { TypeJsonArray } from '@/api/config/types/config-types';
 import { GeoJSON, layerConfigIsGeoJSON } from '@/geo/layer/geoview-layers/vector/geojson';
 import { GeoPackage, layerConfigIsGeoPackage } from '@/geo/layer/geoview-layers/vector/geopackage';
 import { layerConfigIsWMS, WMS } from '@/geo/layer/geoview-layers/raster/wms';
@@ -94,6 +95,7 @@ import { ConfigApi } from '@/api/config/config-api';
 import { GeoViewError } from '@/core/exceptions/geoview-exceptions';
 import { LayerGeoCoreError } from '@/core/exceptions/geocore-exceptions';
 import { ShapefileReader } from '@/core/utils/config/reader/shapefile-reader';
+import { GroupLayerEntryConfig } from '@/core/utils/config/validation-classes/group-layer-entry-config';
 
 /**
  * A class to get the layer from layer type. Layer type can be esriFeature, esriDynamic and ogcWMS
@@ -624,36 +626,8 @@ export class LayerApi {
    * @private
    */
   #addGeoviewLayerStep2(geoviewLayerConfig: TypeGeoviewLayerConfig): GeoViewLayerAddedResult {
-    // TODO: Refactor - Here the function should use the structure created by validation config with the metadata fetch and no need to pass the validation.
-    let layerBeingAdded: AbstractGeoViewLayer;
-    if (layerConfigIsGeoJSON(geoviewLayerConfig)) {
-      layerBeingAdded = new GeoJSON(geoviewLayerConfig);
-    } else if (layerConfigIsGeoPackage(geoviewLayerConfig)) {
-      layerBeingAdded = new GeoPackage(geoviewLayerConfig);
-    } else if (layerConfigIsCSV(geoviewLayerConfig)) {
-      layerBeingAdded = new CSV(geoviewLayerConfig);
-    } else if (layerConfigIsWMS(geoviewLayerConfig)) {
-      layerBeingAdded = new WMS(geoviewLayerConfig, LayerApi.DEBUG_WMS_LAYER_GROUP_FULL_SUB_LAYERS);
-    } else if (layerConfigIsEsriDynamic(geoviewLayerConfig)) {
-      layerBeingAdded = new EsriDynamic(geoviewLayerConfig);
-    } else if (layerConfigIsEsriFeature(geoviewLayerConfig)) {
-      layerBeingAdded = new EsriFeature(geoviewLayerConfig);
-    } else if (layerConfigIsEsriImage(geoviewLayerConfig)) {
-      layerBeingAdded = new EsriImage(geoviewLayerConfig);
-    } else if (layerConfigIsImageStatic(geoviewLayerConfig)) {
-      layerBeingAdded = new ImageStatic(geoviewLayerConfig);
-    } else if (layerConfigIsWFS(geoviewLayerConfig)) {
-      layerBeingAdded = new WFS(geoviewLayerConfig);
-    } else if (layerConfigIsOgcFeature(geoviewLayerConfig)) {
-      layerBeingAdded = new OgcFeature(geoviewLayerConfig);
-    } else if (layerConfigIsXYZTiles(geoviewLayerConfig)) {
-      layerBeingAdded = new XYZTiles(geoviewLayerConfig);
-    } else if (layerConfigIsVectorTiles(geoviewLayerConfig)) {
-      layerBeingAdded = new VectorTiles(geoviewLayerConfig, this.mapViewer.getProjection().getCode());
-    } else {
-      // Not implemented
-      throw new NotSupportedError('Unsupported layer class type');
-    }
+    // Create the layer for the processing
+    const layerBeingAdded = LayerApi.createLayerFromConfigType(geoviewLayerConfig, this.mapViewer.getProjection().getCode());
 
     // Add in the geoviewLayers set
     this.#geoviewLayers[layerBeingAdded.geoviewLayerId] = layerBeingAdded;
@@ -860,8 +834,8 @@ export class LayerApi {
     const geoviewLayer = layerEntryConfig ? this.#geoviewLayers[layerEntryConfig.geoviewLayerConfig.geoviewLayerId] : undefined;
 
     if (layerEntryConfig && geoviewLayer) {
-      if (layerEntryConfig.entryType === 'group') {
-        (layerEntryConfig as unknown as GroupLayerEntryConfig).listOfLayerEntryConfig.forEach((sublayerEntryConfig) => {
+      if (layerEntryConfig instanceof GroupLayerEntryConfig) {
+        layerEntryConfig.listOfLayerEntryConfig.forEach((sublayerEntryConfig) => {
           if ((sublayerEntryConfig as unknown as AbstractBaseLayerEntryConfig).layerStatus === 'error')
             this.reloadLayer((sublayerEntryConfig as unknown as AbstractBaseLayerEntryConfig).layerPath);
         });
@@ -1092,7 +1066,7 @@ export class LayerApi {
 
         // TODO: refactor - remove cast
         if (mapFeaturesConfig.map.listOfGeoviewLayerConfig)
-          mapFeaturesConfig.map.listOfGeoviewLayerConfig = (mapFeaturesConfig.map.listOfGeoviewLayerConfig as MapConfigLayerEntry[]).filter(
+          mapFeaturesConfig.map.listOfGeoviewLayerConfig = mapFeaturesConfig.map.listOfGeoviewLayerConfig.filter(
             (geoviewLayerConfig) => geoviewLayerConfig.geoviewLayerId !== layerPath
           );
       } else if (layerPathNodes.length === 2) {
@@ -2396,7 +2370,362 @@ export class LayerApi {
     return newOrderedLayerInfos;
   }
 
+  static createLayerFromConfigType(geoviewLayerConfig: TypeGeoviewLayerConfig, mapProjectionForVectorTiles: string): AbstractGeoViewLayer {
+    // TODO: Refactor - Here the function should use the structure created by validation config with the metadata fetch and no need to pass the validation.
+    if (layerConfigIsGeoJSON(geoviewLayerConfig)) {
+      return new GeoJSON(geoviewLayerConfig);
+    }
+    if (layerConfigIsGeoPackage(geoviewLayerConfig)) {
+      return new GeoPackage(geoviewLayerConfig);
+    }
+    if (layerConfigIsCSV(geoviewLayerConfig)) {
+      return new CSV(geoviewLayerConfig);
+    }
+    if (layerConfigIsWMS(geoviewLayerConfig)) {
+      return new WMS(geoviewLayerConfig, LayerApi.DEBUG_WMS_LAYER_GROUP_FULL_SUB_LAYERS);
+    }
+    if (layerConfigIsEsriDynamic(geoviewLayerConfig)) {
+      return new EsriDynamic(geoviewLayerConfig);
+    }
+    if (layerConfigIsEsriFeature(geoviewLayerConfig)) {
+      return new EsriFeature(geoviewLayerConfig);
+    }
+    if (layerConfigIsEsriImage(geoviewLayerConfig)) {
+      return new EsriImage(geoviewLayerConfig);
+    }
+    if (layerConfigIsImageStatic(geoviewLayerConfig)) {
+      return new ImageStatic(geoviewLayerConfig);
+    }
+    if (layerConfigIsWFS(geoviewLayerConfig)) {
+      return new WFS(geoviewLayerConfig);
+    }
+    if (layerConfigIsOgcFeature(geoviewLayerConfig)) {
+      return new OgcFeature(geoviewLayerConfig);
+    }
+    if (layerConfigIsXYZTiles(geoviewLayerConfig)) {
+      return new XYZTiles(geoviewLayerConfig);
+    }
+    if (layerConfigIsVectorTiles(geoviewLayerConfig)) {
+      return new VectorTiles(geoviewLayerConfig, mapProjectionForVectorTiles);
+    }
+
+    // Not implemented
+    throw new NotSupportedError('Unsupported layer class type');
+  }
+
   // #endregion
+
+  // #region Experimental
+
+  /**
+   * Experimental approach to use our Geoview-Layers classes from the ConfigAPI
+   * @returns A Promise with the layer configuration
+   * @experimental
+   */
+  // TODO: REFACTOR CONFIG API
+  static createEsriDynamicConfig(): Promise<ConfigBaseClass[]> {
+    // Create the Layer config
+    const layerConfig = EsriDynamic.createEsriDynamicLayerConfig(
+      'gvLayerID',
+      'Some Name',
+      'https://maps-cartes.ec.gc.ca/arcgis/rest/services/CESI/MapServer',
+      false,
+      [{ index: '4' }] as unknown as TypeJsonArray,
+      {}
+    );
+
+    // Create the class from geoview-layers package
+    const myLayer = new EsriDynamic(layerConfig);
+
+    // Process it
+    return LayerApi.#processConfig(myLayer);
+  }
+
+  /**
+   * Experimental approach to use our Geoview-Layers classes from the ConfigAPI
+   * @returns A Promise with the layer configuration
+   * @experimental
+   */
+  // TODO: REFACTOR CONFIG API
+  static createEsriImageConfig(): Promise<ConfigBaseClass[]> {
+    // Create the Layer config
+    const layerConfig = EsriImage.createEsriImageLayerConfig(
+      'gvLayerID',
+      'Some Name',
+      'https://www5.agr.gc.ca/atlas/rest/services/app_agclimate_agclimat/agclimate_tx/ImageServer',
+      false
+    );
+
+    // Create the class from geoview-layers package
+    const myLayer = new EsriImage(layerConfig);
+
+    // Process it
+    return LayerApi.#processConfig(myLayer);
+  }
+
+  /**
+   * Experimental approach to use our Geoview-Layers classes from the ConfigAPI
+   * @returns A Promise with the layer configuration
+   * @experimental
+   */
+  // TODO: REFACTOR CONFIG API
+  static createImageStaticConfig(): Promise<ConfigBaseClass[]> {
+    // Create the Layer config
+    const layerConfig = ImageStatic.createImageStaticLayerConfig(
+      'gvLayerID',
+      'Some Name',
+      'https://datacube-prod-data-public.s3.ca-central-1.amazonaws.com/store/imagery/aerial/napl/napl-ring-of-fire',
+      false,
+      [{ id: 'napl-ring-of-fire-1954-08-07-60k-thumbnail.png' }] as unknown as TypeJsonArray
+    );
+
+    // Create the class from geoview-layers package
+    const myLayer = new ImageStatic(layerConfig);
+
+    // Process it
+    return LayerApi.#processConfig(myLayer);
+  }
+
+  /**
+   * Experimental approach to use our Geoview-Layers classes from the ConfigAPI
+   * @returns A Promise with the layer configuration
+   * @experimental
+   */
+  // TODO: REFACTOR CONFIG API
+  static createVectorTilesConfig(): Promise<ConfigBaseClass[]> {
+    // Create the Layer config
+    const layerConfig = VectorTiles.createVectorTilesLayerConfig(
+      'gvLayerID',
+      'Some Name',
+      'https://tiles.arcgis.com/tiles/HsjBaDykC1mjhXz9/arcgis/rest/services/CBMT_CBCT_3978_V_OSM/VectorTileServer',
+      false,
+      [{ id: 'CBMT_CBCT_3978_V_OSM' }] as unknown as TypeJsonArray
+    );
+
+    // Create the class from geoview-layers package
+    const myLayer = new VectorTiles(layerConfig, 'EPSG:3978');
+
+    // Process it
+    return LayerApi.#processConfig(myLayer);
+  }
+
+  /**
+   * Experimental approach to use our Geoview-Layers classes from the ConfigAPI
+   * @returns A Promise with the layer configuration
+   * @experimental
+   */
+  // TODO: REFACTOR CONFIG API
+  static createWMSConfig(): Promise<ConfigBaseClass[]> {
+    // Create the Layer config
+    const layerConfig = WMS.createWMSLayerConfig(
+      'gvLayerID',
+      'Some Name',
+      'https://maps-cartes.services.geo.ca/server_serveur/services/NRCan/CER_Assessments_EN/MapServer/WMSServer',
+      'mapserver',
+      false,
+      [{ id: '0' }] as unknown as TypeJsonArray,
+      {}
+    );
+
+    // Create the class from geoview-layers package
+    const myLayer = new WMS(layerConfig, false);
+
+    // Process it
+    return LayerApi.#processConfig(myLayer);
+  }
+
+  /**
+   * Experimental approach to use our Geoview-Layers classes from the ConfigAPI
+   * @returns A Promise with the layer configuration
+   * @experimental
+   */
+  // TODO: REFACTOR CONFIG API
+  static createXYZTilesConfig(): Promise<ConfigBaseClass[]> {
+    // Create the Layer config
+    const layerConfig = XYZTiles.createXYZTilesLayerConfig(
+      'gvLayerID',
+      'Some Name',
+      'https://server.arcgisonline.com/ArcGIS/rest/services/World_Topo_Map/MapServer',
+      false,
+      [{ id: '0' }] as unknown as TypeJsonArray
+    );
+
+    // Create the class from geoview-layers package
+    const myLayer = new XYZTiles(layerConfig);
+
+    // Process it
+    return LayerApi.#processConfig(myLayer);
+  }
+
+  /**
+   * Experimental approach to use our Geoview-Layers classes from the ConfigAPI
+   * @returns A Promise with the layer configuration
+   * @experimental
+   */
+  // TODO: REFACTOR CONFIG API
+  static createCSVFeatureConfig(): Promise<ConfigBaseClass[]> {
+    // Create the Layer config
+    const layerConfig = CSV.createCSVLayerConfig(
+      'gvLayerID',
+      'Some Name',
+      'https://canadian-geospatial-platform.github.io/geoview/public/datasets/csv-files',
+      false,
+      [{ id: 'NPRI-INRP_WaterEau_MediaGroupMilieu_2022' }] as unknown as TypeJsonArray
+    );
+
+    // Create the class from geoview-layers package
+    const myLayer = new CSV(layerConfig);
+
+    // Process it
+    return LayerApi.#processConfig(myLayer);
+  }
+
+  /**
+   * Experimental approach to use our Geoview-Layers classes from the ConfigAPI
+   * @returns A Promise with the layer configuration
+   * @experimental
+   */
+  // TODO: REFACTOR CONFIG API
+  static createEsriFeatureConfig(): Promise<ConfigBaseClass[]> {
+    // Create the Layer config
+    const layerConfig = EsriFeature.createEsriFeatureLayerConfig(
+      'gvLayerID',
+      'Some Name',
+      'https://services.arcgis.com/V6ZHFr6zdgNZuVG0/ArcGIS/rest/services/Toronto_Neighbourhoods/FeatureServer',
+      false,
+      [{ index: '0' }] as unknown as TypeJsonArray
+    );
+
+    // Create the class from geoview-layers package
+    const myLayer = new EsriFeature(layerConfig);
+
+    // Process it
+    return LayerApi.#processConfig(myLayer);
+  }
+
+  /**
+   * Experimental approach to use our Geoview-Layers classes from the ConfigAPI
+   * @returns A Promise with the layer configuration
+   * @experimental
+   */
+  // TODO: REFACTOR CONFIG API
+  static createGeoJsonLayerConfig(): Promise<ConfigBaseClass[]> {
+    // Create the Layer config
+    const layerConfig = GeoJSON.createGeoJsonLayerConfig(
+      'gvLayerID',
+      'Some Name',
+      'https://canadian-geospatial-platform.github.io/geoview/public/datasets/geojson/metadata.meta',
+      false,
+      [{ id: 'polygons.json' }] as unknown as TypeJsonArray
+    );
+
+    // Create the class from geoview-layers package
+    const myLayer = new GeoJSON(layerConfig);
+
+    // Process it
+    return LayerApi.#processConfig(myLayer);
+  }
+
+  /**
+   * Experimental approach to use our Geoview-Layers classes from the ConfigAPI
+   * @returns A Promise with the layer configuration
+   * @experimental
+   */
+  // TODO: REFACTOR CONFIG API
+  static createOGCFeatureConfig(): Promise<ConfigBaseClass[]> {
+    // Create the Layer config
+    const layerConfig = OgcFeature.createOgcFeatureLayerConfig(
+      'gvLayerID',
+      'Some Name',
+      'https://b6ryuvakk5.execute-api.us-east-1.amazonaws.com/dev',
+      false,
+      [{ id: 'lakes' }] as unknown as TypeJsonArray
+    );
+
+    // Create the class from geoview-layers package
+    const myLayer = new OgcFeature(layerConfig);
+
+    // Process it
+    return LayerApi.#processConfig(myLayer);
+  }
+
+  /**
+   * Experimental approach to use our Geoview-Layers classes from the ConfigAPI
+   * @returns A Promise with the layer configuration
+   * @experimental
+   */
+  // TODO: REFACTOR CONFIG API
+  static createWFSConfig(): Promise<ConfigBaseClass[]> {
+    // Create the Layer config
+    const layerConfig = WFS.createWfsFeatureLayerConfig(
+      'gvLayerID',
+      'Some Name',
+      'https://ahocevar.com/geoserver/wfs?REQUEST=GetCapabilities&VERSION=2.0.0&SERVICE=WFS',
+      false,
+      [{ id: 'usa:states' }] as unknown as TypeJsonArray
+    );
+
+    // Create the class from geoview-layers package
+    const myLayer = new WFS(layerConfig);
+
+    // Process it
+    return LayerApi.#processConfig(myLayer);
+  }
+
+  /**
+   * Processes a Layer Config by calling 'createGeoViewLayers' on the provided layer.
+   * @param {AbstractGeoViewLayer} layer - The layer to use to process the configuration
+   * @returns {Promise<ConfigBaseClass>} The promise of a generated ConfigBaseClass.
+   * @private
+   */
+  static #processConfig(layer: AbstractGeoViewLayer): Promise<ConfigBaseClass[]> {
+    // Create a promise that the layer config will be created
+    const promise = new Promise<ConfigBaseClass[]>((resolve, reject) => {
+      // Register a handler when the layer config has been created for this config
+      layer.onLayerConfigCreated((geoviewLayer: AbstractGeoViewLayer, event: LayerConfigCreatedEvent) => {
+        // A Layer Config was created
+        logger.logDebug('Config created', event.config);
+      });
+
+      // (Extra) Register a handler when a Group layer has been created for this config
+      layer.onLayerGroupCreated((geoviewLayer: AbstractGeoViewLayer, event: LayerGroupCreatedEvent) => {
+        // A Group Layer was created
+        logger.logDebug('Group Layer created', event.layer);
+      });
+
+      // (Extra) Register a handler when a GV layer has been created for this config
+      layer.onLayerGVCreated((geoviewLayer: AbstractGeoViewLayer, event: LayerGVCreatedEvent) => {
+        // A GV Layer was created
+        logger.logDebug('GV Layer created', event.layer);
+      });
+
+      // Start the geoview-layers config process
+      layer
+        .createGeoViewLayers()
+        .then((configs) => {
+          // If no errors
+          if (layer.layerLoadError.length === 0) {
+            // Resolve with the configurations
+            resolve(configs);
+          } else {
+            // Errors happened
+            // If only one, throw as-is
+            if (layer.layerLoadError.length === 1) throw layer.layerLoadError[0];
+            // Throw them all inside an AggregateError
+            layer.throwAggregatedLayerLoadErrors();
+          }
+        })
+        .catch((error: unknown) => {
+          // Reject
+          reject(formatError(error));
+        });
+    });
+
+    // Return the promise
+    return promise;
+  }
+
+  // #endregion Experimental
 }
 
 export type GeoViewLayerAddedResult = {
